third_party/blink/web_tests/platform/generic/fast/events/onchange-range-slider-expected.txt

This test verifies that updating the slider for an input element with type=range fires a change event.

PASS successfullyParsed is true

TEST COMPLETE

PASS Change event fired.
PASS slider.value is "0"
